refactor(web): osk key class defs -> separate files
