+++ web/source/osk/visualKeyboard.ts
@@ -1,576 +1,11 @@
 /// <reference path="preProcessor.ts" />
 /// <reference path="utils.ts" />
+/// <reference path="oskBaseKey.ts" />
 /// <reference path="keytip.interface.ts" />
 /// <reference path="browser/keytip.ts" />
 /// <reference path="browser/subkeyPopup.ts" />
 
 namespace com.keyman.osk {
-  let Codes = com.keyman.text.Codes;
-  //#region Definition of the KeyElement merger type
-  class KeyData {
-    ['key']: OSKKey;
-    ['keyId']: string;
-    ['subKeys']?: OSKKeySpec[];
-
-    constructor(keyData: OSKKey, keyId: string) {
-      this['key'] = keyData;
-      this['keyId'] = keyId;
-    }
-  }
-
-  export type KeyElement = HTMLDivElement & KeyData;
-
-  // Many thanks to https://www.typescriptlang.org/docs/handbook/advanced-types.html for this.
-  function link(elem: HTMLDivElement, data: KeyData): KeyElement {
-    let e = <KeyElement> elem;
-
-    // Merges all properties and methods of KeyData onto the underlying HTMLDivElement, creating a merged class.
-    for(let id in data) {
-      if(!e.hasOwnProperty(id)) {
-        (<any>e)[id] = (<any>data)[id];
-      }
-    }
-
-    return e;
-  }
-
-  export function isKey(elem: Node): boolean {
-    return elem && ('key' in elem) && ((<any> elem['key']) instanceof OSKKey);
-  }
-
-  export function getKeyFrom(elem: Node): KeyElement {
-    if(isKey(elem)) {
-      return <KeyElement> elem;
-    } else {
-      return null;
-    }
-  }
-  //#endregion
-
-  //#region OSK key objects and construction
-  export class OSKKeySpec implements keyboards.LayoutKey {
-    id: string;
-
-    // Only set (within @keymanapp/keyboard-processor) for keys actually specified in a loaded layout
-    baseKeyID?: string;
-    coreID?: string;
-    elementID?: string;
-
-    text?: string;
-    sp?: number | keyboards.ButtonClass;
-    width: string;
-    layer?: string; // The key will derive its base modifiers from this property - may not equal the layer on which it is displayed.
-    nextlayer?: string;
-    pad?: string;
-    widthpc?: number; // Added during OSK construction.
-    padpc?: number; // Added during OSK construction.
-    sk?: OSKKeySpec[];
-
-    constructor(id: string, text?: string, width?: string, sp?: number | keyboards.ButtonClass, nextlayer?: string, pad?: string) {
-      this.id = id;
-      this.text = text;
-      this.width = width ? width : "50";
-      this.sp = sp;
-      this.nextlayer = nextlayer;
-      this.pad = pad;
-    }
-  }
-
-  export abstract class OSKKey {
-    spec: OSKKeySpec;
-    btn: KeyElement;
-    label: HTMLSpanElement;
-
-    /**
-     * The layer of the OSK on which the key is displayed.
-     */
-    readonly layer: string;
-
-    constructor(spec: OSKKeySpec, layer: string) {
-      this.spec = spec;
-      this.layer = layer;
-    }
-
-    abstract getId(): string;
-
-    /**
-     * Attach appropriate class to each key button, according to the layout
-     *
-     * @param       {Object=}   layout  source layout description (optional, sometimes)
-     */
-    public setButtonClass(vkbd: VisualKeyboard) {
-      let key = this.spec;
-      let btn = this.btn;
-
-      var n=0, keyTypes=['default','shift','shift-on','special','special-on','','','','deadkey','blank','hidden'];
-      if(typeof key['dk'] == 'string' && key['dk'] == '1') {
-        n=8;
-      }
-
-      if(typeof key['sp'] == 'string') {
-        n=parseInt(key['sp'],10);
-      }
-
-      if(n < 0 || n > 10) {
-        n=0;
-      }
-
-      // Apply an overriding class for 5-row layouts
-      var nRows=vkbd.layout['layer'][0]['row'].length;
-      if(nRows > 4 && vkbd.device.formFactor == 'phone') {
-        btn.className='kmw-key kmw-5rows kmw-key-'+keyTypes[n];
-      } else {
-        btn.className='kmw-key kmw-key-'+keyTypes[n];
-      }
-    }
-
-    /**
-     * Uses canvas.measureText to compute and return the width of the given text of given font in pixels.
-     *
-     * @param {String} text The text to be rendered.
-     * @param {String} style The CSSStyleDeclaration for an element to measure against, without modification.
-     *
-     * @see https://stackoverflow.com/questions/118241/calculate-text-width-with-javascript/21015393#21015393
-     * This version has been substantially modified to work for this particular application.
-     */
-    static getTextMetrics(text: string, emScale: number, style: {fontFamily?: string, fontSize: string}): TextMetrics {
-      // A final fallback - having the right font selected makes a world of difference.
-      if(!style.fontFamily) {
-        style.fontFamily = getComputedStyle(document.body).fontFamily;
-      }
-
-      if(!style.fontSize || style.fontSize == "") {
-        style.fontSize = '1em';
-      }
-
-      let fontFamily = style.fontFamily;
-      let fontSpec = getFontSizeStyle(style.fontSize);
-
-      var fontSize: string;
-      if(fontSpec.absolute) {
-        // We've already got an exact size - use it!
-        fontSize = fontSpec.val + 'px';
-      } else {
-        fontSize = fontSpec.val * emScale + 'px';
-      }
-
-      // re-use canvas object for better performance
-      var canvas: HTMLCanvasElement = OSKKey.getTextMetrics['canvas'] || 
-                                     (OSKKey.getTextMetrics['canvas'] = document.createElement("canvas"));
-      var context = canvas.getContext("2d");
-      context.font = fontSize + " " + fontFamily;
-      var metrics = context.measureText(text);
-
-      return metrics;
-    }
-
-    getIdealFontSize(osk: VisualKeyboard, style: {height?: string, fontFamily?: string, fontSize: string}): string {
-      // Recompute the new width for use in autoscaling calculations below, just in case.
-      let emScale = osk.getKeyEmFontSize();
-      let metrics = OSKKey.getTextMetrics(this.spec.text, emScale, style);
-
-      let fontSpec = getFontSizeStyle(style.fontSize);
-      let keyWidth = this.getKeyWidth(osk);
-      const MAX_X_PROPORTION = 0.90;
-      const MAX_Y_PROPORTION = 0.90;
-      const X_PADDING = 2;
-      const Y_PADDING = 2;
-
-      var fontHeight: number, keyHeight: number;
-      if(metrics.fontBoundingBoxAscent) {
-        fontHeight = metrics.fontBoundingBoxAscent + metrics.fontBoundingBoxDescent;
-      }
-
-      let textHeight = fontHeight ? fontHeight + Y_PADDING : 0;
-      if(style.height && style.height.indexOf('px') != -1) {
-        keyHeight = Number.parseFloat(style.height.substring(0, style.height.indexOf('px')));
-      }
-
-      let xProportion = (keyWidth * MAX_X_PROPORTION) / (metrics.width + X_PADDING); // How much of the key does the text want to take?
-      let yProportion = textHeight && keyHeight ? (keyHeight * MAX_Y_PROPORTION) / textHeight : undefined;
-
-      var proportion: number = xProportion;
-      if(yProportion && yProportion < xProportion) {
-        proportion = yProportion;
-      }
-
-      // Never upscale keys past the default - only downscale them.
-      // Proportion < 1:  ratio of key width to (padded [loosely speaking]) text width
-      //                  maxProportion determines the 'padding' involved.
-      if(proportion < 1) {
-        if(fontSpec.absolute) {
-          return proportion * fontSpec.val + 'px';
-        } else {
-          return proportion * fontSpec.val + 'em';
-        }
-      }
-    }
-
-    getKeyWidth(vkbd: VisualKeyboard): number {
-      let units = this.objectUnits(vkbd);
-
-      if(units == 'px') {
-        // For mobile devices, we presently specify width directly in pixels.  Just use that!
-        return this.spec['widthpc'];
-      } else if(units == '%') {
-        // For desktop devices, each key is given a %age of the total OSK width.  We'll need to compute an
-        // approximation for that.  `this.kbdDiv` is the element controlling the OSK's width, set in px.
-
-        // This is an approximation that tends to be a bit too large, but it's close enough to be useful.
-        return Math.floor(vkbd.width * this.spec['widthpc'] / 100);
-      }
-    }
-
-    objectUnits(vkbd: VisualKeyboard): string {
-      // Returns a unit string corresponding to how the width for each key is specified.
-      if(vkbd.device.formFactor == 'desktop' || vkbd.isStatic) {
-        return '%';
-      } else {
-        return 'px';
-      }
-    }
-
-    /**
-     * Replace default key names by special font codes for modifier keys
-     *
-     *  @param  {string}  oldText
-     *  @return {string}
-     **/
-    protected renameSpecialKey(oldText: string, osk: VisualKeyboard): string {
-      // If a 'special key' mapping exists for the text, replace it with its corresponding special OSK character.
-      switch(oldText) {
-        case '*ZWNJ*':
-          // Default ZWNJ symbol comes from iOS.  We'd rather match the system defaults where
-          // possible / available though, and there's a different standard symbol on Android.
-          oldText = osk.device.coreSpec.OS == com.keyman.utils.OperatingSystem.Android ?
-            '*ZWNJAndroid*' :
-            '*ZWNJiOS*';
-          break;
-        case '*Enter*':
-          oldText = osk.isRTL ? '*RTLEnter*' : '*LTREnter*';
-          break;
-        case '*BkSp*':
-          oldText = osk.isRTL ? '*RTLBkSp*' : '*LTRBkSp*';
-          break;
-        default:
-          // do nothing.
-      }
-
-      let specialCodePUA = 0XE000 + VisualKeyboard.specialCharacters[oldText];
-
-      return VisualKeyboard.specialCharacters[oldText] ?
-        String.fromCharCode(specialCodePUA) :
-        oldText;
-    }
-
-    // Produces a HTMLSpanElement with the key's actual text.
-    protected generateKeyText(osk: VisualKeyboard): HTMLSpanElement {
-      let spec = this.spec;
-
-      // Add OSK key labels
-      var keyText;
-      var t = document.createElement('span'), ts=t.style;
-      if(spec['text'] == null || spec['text'] == '') {
-        keyText='\xa0';  // default:  nbsp.
-        if(typeof spec['id'] == 'string') {
-          // If the ID's Unicode-based, just use that code.
-          if(/^U_[0-9A-F]{4}$/i.test(spec['id'])) {
-            keyText=String.fromCharCode(parseInt(spec['id'].substr(2),16));
-          }
-        }
-      } else {
-        keyText=spec['text'];
-
-        // Unique layer-based transformation:  SHIFT-TAB uses a different glyph.
-        if(keyText == '*Tab*' && this.layer == 'shift') {
-          keyText = '*TabLeft*';
-        }
-      }
-
-      t.className='kmw-key-text';
-
-      let specialText = this.renameSpecialKey(keyText, osk);
-      if(specialText != keyText) {
-        // The keyboard wants to use the code for a special glyph defined by the SpecialOSK font.
-        keyText = specialText;
-        spec['font'] = "SpecialOSK";
-      }
-
-      // Grab our default for the key's font and font size.
-      ts.fontSize=osk.fontSize;     //Build 344, KMEW-90
-
-      //Override font spec if set for this key in the layout
-      if(typeof spec['font'] == 'string' && spec['font'] != '') {
-        ts.fontFamily=spec['font'];
-      }
-
-      if(typeof spec['fontsize'] == 'string' && spec['fontsize'] != '') {
-        ts.fontSize=spec['fontsize'];
-      }
-
-      // For some reason, fonts will sometimes 'bug out' for the embedded iOS page if we
-      // instead assign fontFamily to the existing style 'ts'.  (Occurs in iOS 12.)
-      let styleSpec: {fontFamily?: string, fontSize: string} = {fontSize: ts.fontSize};
-
-      if(ts.fontFamily) {
-        styleSpec.fontFamily = ts.fontFamily;
-      } else {
-        styleSpec.fontFamily = osk.fontFamily; // Helps with style sheet calculations.
-      }
-
-      // Check the key's display width - does the key visualize well?
-      let emScale = osk.getKeyEmFontSize();
-      var width: number = OSKKey.getTextMetrics(keyText, emScale, styleSpec).width;
-      if(width == 0 && keyText != '' && keyText != '\xa0') {
-        // Add the Unicode 'empty circle' as a base support for needy diacritics.
-
-        // Disabled by mcdurdin 2020-10-19; dotted circle display is inconsistent on iOS/Safari
-        // at least and doesn't combine with diacritic marks. For consistent display, it may be
-        // necessary to build a custom font that does not depend on renderer choices for base
-        // mark display -- e.g. create marks with custom base included, potentially even on PUA
-        // code points and use those in rendering the OSK. See #3039 for more details.
-        // keyText = '\u25cc' + keyText;
-
-        if(osk.isRTL) {
-          // Add the RTL marker to ensure it displays properly.
-          keyText = '\u200f' + keyText;
-        }
-      }
-
-      ts.fontSize = this.getIdealFontSize(osk, styleSpec);
-
-      // Finalize the key's text.
-      t.innerHTML = keyText;
-
-      return t;
-    }
-  }
-
-  export class OSKBaseKey extends OSKKey {
-    constructor(spec: OSKKeySpec, layer: string) {
-      super(spec, layer);
-    }
-
-    getId(): string {
-      // Define each key element id by layer id and key id (duplicate possible for SHIFT - does it matter?)
-      return this.spec.elementID;
-    }
-
-    // Produces a small reference label for the corresponding physical key on a US keyboard.
-    private generateKeyCapLabel(): HTMLDivElement {
-      // Create the default key cap labels (letter keys, etc.)
-      var x = Codes.keyCodes[this.spec.baseKeyID];
-      switch(x) {
-        // Converts the keyman key id code for common symbol keys into its representative ASCII code.
-        // K_COLON -> K_BKQUOTE
-        case 186: x=59; break;
-        case 187: x=61; break;
-        case 188: x=44; break;
-        case 189: x=45; break;
-        case 190: x=46; break;
-        case 191: x=47; break;
-        case 192: x=96; break;
-        // K_LBRKT -> K_QUOTE
-        case 219: x=91; break;
-        case 220: x=92; break;
-        case 221: x=93; break;
-        case 222: x=39; break;
-        default:
-          // No other symbol character represents a base key on the standard QWERTY English layout.
-          if(x < 48 || x > 90) {
-            x=0;
-          }
-      }
-
-      if(x > 0) {
-        let q = document.createElement('div');
-        q.className='kmw-key-label';
-        q.innerHTML=String.fromCharCode(x);
-        return q;
-      } else {
-        // Keyman-only virtual keys have no corresponding physical key.
-        return null;
-      }
-    }
-
-    private processSubkeys(btn: KeyElement, osk: VisualKeyboard) {
-      // Add reference to subkey array if defined
-      var bsn: number, bsk=btn['subKeys'] = this.spec['sk'];
-      // Transform any special keys into their PUA representations.
-      for(bsn=0; bsn<bsk.length; bsn++) {
-        if(bsk[bsn]['sp'] == '1' || bsk[bsn]['sp'] == '2') {
-          var oldText=bsk[bsn]['text'];
-          bsk[bsn]['text']=this.renameSpecialKey(oldText, osk);
-        }
-
-        // If a subkey doesn't have a defined layer property, copy it from the base key's layer by default.
-        if(!bsk[bsn].layer) {
-          bsk[bsn].layer = btn.key.layer
-        }
-      }
-
-      // If a subkey array is defined, add an icon
-      var skIcon = document.createElement('div');
-      skIcon.className='kmw-key-popup-icon';
-      //kDiv.appendChild(skIcon);
-      btn.appendChild(skIcon);
-    }
-
-    construct(osk: VisualKeyboard, layout: keyboards.LayoutFormFactor, rowStyle: CSSStyleDeclaration, totalPercent: number): {element: HTMLDivElement, percent: number} {
-      let spec = this.spec;
-      let isDesktop = osk.device.formFactor == "desktop"
-
-      let kDiv = document.createElement('div');
-      kDiv.className='kmw-key-square';
-
-      let ks=kDiv.style;
-      ks.width=this.objectGeometry(osk, spec['widthpc']);
-
-      let originalPercent = totalPercent;
-
-      let btnEle = document.createElement('div');
-      let btn = this.btn = link(btnEle, new KeyData(this, spec['id']));
-
-      // Set button class
-      this.setButtonClass(osk);
-
-      // Set key and button positioning properties.
-      if(!isDesktop) {
-        // Regularize interkey spacing by rounding key width and padding (Build 390)
-        ks.left=this.objectGeometry(osk, totalPercent+spec['padpc']);
-        if(!osk.isStatic) {
-          ks.bottom=rowStyle.bottom;
-        }
-        ks.height=rowStyle.height;  // must be specified in px for rest of layout to work correctly
-
-        if(!osk.isStatic) {
-          // Set distinct phone and tablet button position properties
-          btn.style.left=ks.left;
-          btn.style.width=ks.width;
-        }
-      } else {
-        ks.marginLeft=this.objectGeometry(osk, spec['padpc']);
-      }
-
-      totalPercent=totalPercent+spec['padpc']+spec['widthpc'];
-
-      // Add the (US English) keycap label for layouts requesting display of underlying keys
-      if(layout["displayUnderlying"]) {
-        let keyCap = this.generateKeyCapLabel();
-
-        if(keyCap) {
-          btn.appendChild(keyCap);
-        }
-      }
-
-      // Define each key element id by layer id and key id (duplicate possible for SHIFT - does it matter?)
-      btn.id=this.getId();
-
-      // Define callbacks to handle key touches: iOS and Android tablets and phones
-      // TODO: replace inline function calls??
-      if(!osk.isStatic && !osk.device.touchable) {
-        // Highlight key while mouse down or if moving back over originally selected key
-        btn.onmouseover=btn.onmousedown=osk.mouseOverMouseDownHandler; // Build 360
-
-        // Remove highlighting when key released or moving off selected element
-        btn.onmouseup=btn.onmouseout=osk.mouseUpMouseOutHandler; //Build 360
-      }
-
-      // Make sure the key text is the element's first child - processSubkeys()
-      // will add an extra element if subkeys exist, which can interfere with
-      // keyboard/language name display on the space bar!
-      btn.appendChild(this.label = this.generateKeyText(osk));
-
-      // Handle subkey-related tasks.
-      if(typeof(spec['sk']) != 'undefined' && spec['sk'] != null) {
-        this.processSubkeys(btn, osk);
-      } else {
-        btn['subKeys']=null;
-      }
-
-      // Add text to button and button to placeholder div
-      kDiv.appendChild(btn);
-
-      // Prevent user selection of key captions
-      //t.style.webkitUserSelect='none';
-
-      // The 'return value' of this process.
-      return {element: kDiv, percent: totalPercent - originalPercent};
-    }
-
-    objectGeometry(vkbd: VisualKeyboard, v: number): string {
-      let unit = this.objectUnits(vkbd);
-      if(unit == '%') {
-        return v + unit;
-      } else { // unit == 'px'
-        return (Math.round(v*100)/100)+unit; // round to 2 decimal places, making css more readable
-      }
-    }
-  }
-
-  export class OSKSubKey extends OSKKey {
-    constructor(spec: OSKKeySpec, layer: string) {
-      if(typeof(layer) != 'string' || layer == '') {
-        throw "The 'layer' parameter for subkey construction must be properly defined.";
-      }
-
-      super(spec, layer);
-    }
-
-    getId(): string {
-      // Create (temporarily) unique ID by prefixing 'popup-' to actual key ID
-      return 'popup-'+this.layer+'-'+this.spec['id'];
-    }
-
-    construct(osk: VisualKeyboard, baseKey: KeyElement, topMargin: boolean): HTMLDivElement {
-      let spec = this.spec;
-
-      let kDiv=document.createElement('div');
-      let tKey = osk.getDefaultKeyObject();
-      let ks=kDiv.style;
-
-      for(var tp in tKey) {
-        if(typeof spec[tp] != 'string') {
-          spec[tp]=tKey[tp];
-        }
-      }
-
-      kDiv.className='kmw-key-square-ex';
-      if(topMargin) {
-        ks.marginTop='5px';
-      }
-
-      if(typeof spec['width'] != 'undefined') {
-        ks.width=(parseInt(spec['width'],10)*baseKey.offsetWidth/100)+'px';
-      } else {
-        ks.width=baseKey.offsetWidth+'px';
-      }
-      ks.height=baseKey.offsetHeight+'px';
-
-      let btnEle=document.createElement('div');
-      let btn = this.btn = link(btnEle, new KeyData(this, spec['id']));
-
-      this.setButtonClass(osk);
-      btn.id = this.getId();
-
-      // Must set button size (in px) dynamically, not from CSS
-      let bs=btn.style;
-      bs.height=ks.height;
-      bs.lineHeight=baseKey.style.lineHeight;
-      bs.width=ks.width;
-
-      // Must set position explicitly, at least for Android
-      bs.position='absolute';
-
-      btn.appendChild(this.label = this.generateKeyText(osk));
-      kDiv.appendChild(btn);
-
-      return kDiv;
-    }
-  }
-
-  //#endregion
-
   export class VisualKeyboard {
     // Defines the PUA code mapping for the various 'special' modifier/control keys on keyboards.
     // `specialCharacters` must be kept in sync with the same variable in builder.js. See also CompileKeymanWeb.pas: CSpecialText10
